fix i0 rename .gitignore file to exclude streamlit api key
Changes the name of the gitignore file to the form that git accepts.
This ensures that github (and any other forks) no longer receive
the original openai api key accidentally committed to this repository.

This commit does not remove the original file, as that will probably
surprise the originating repository committer. Hopefully, that person
will use `git rm -r -cached .streamlib` or some equivalent, per
https://stackoverflow.com/q/1274057/22365152 .